use unsigned int make col_import_array_len unsigned
2004-10-21  Andreas J. Guelzow <aguelzow@taliesin.ca>

	* src/stf.c: (stf_apply_formats): use unsigned int
	* src/stf-parse.h: make col_import_array_len unsigned

+++ src/stf.c
@@ -136,7 +136,7 @@ stf_apply_formats (StfParseOptions_t *parseoptions,
 
 	for (ui = 0; ui < parseoptions->formats->len; ui++) {
 		if (parseoptions->col_import_array == NULL ||
-		    parseoptions->col_import_array_len <= (int)ui ||
+		    parseoptions->col_import_array_len <= ui ||
 		    parseoptions->col_import_array[ui]) {
 			GnmStyle *style = mstyle_new ();
 			GnmFormat *sf = g_ptr_array_index
